backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.0.350
openai==1.3.5
chromadb==0.4.18
pypdf==3.17.1
tiktoken==0.5.1
python-dotenv==1.0.0